+++ papa-dev/src/main/java/papa/internal/ForceShellProfileableReceiver.kt
@@ -5,8 +5,8 @@ import android.content.Context
 import android.content.Intent
 import papa.SafeTrace
 
-internal class ForceTraceableReceiver : BroadcastReceiver() {
+internal class ForceShellProfileableReceiver : BroadcastReceiver() {
   override fun onReceive(context: Context, intent: Intent) {
-    SafeTrace.forceTraceable()
+    SafeTrace.forceShellProfileable()
   }
 }